+++ .storybook/preview.ts
@@ -1,5 +1,5 @@
 import type { Preview } from "@storybook/react";
-import customTheme from "../src/lib/styles/theme";
+import defaultTheme from "../src/lib/styles/theme";
 
 const preview: Preview = {
   parameters: {
@@ -11,7 +11,7 @@ const preview: Preview = {
       },
     },
     chakra: {
-      theme: customTheme,
+      theme: defaultTheme,
     },
   },
 };