+++ src/lib/services/expression/index.ts
@@ -1,2 +1,1 @@
 export * from "./poolExpression";
-export * from "./txExpression";